REF: Move relevant emulators into twined
