+++ octue/twined/cloud/emulators/child.py
@@ -2,7 +2,6 @@ import logging
 
 from octue.cloud.events.replayer import EventReplayer
 
-
 logger = logging.getLogger(__name__)
