+++ tests/cloud/pub_sub/test_logging.py
@@ -3,11 +3,11 @@ import logging
 from logging import makeLogRecord
 from unittest.mock import patch
 
-from octue.cloud.emulators._pub_sub import MESSAGES, MockService
-from octue.cloud.emulators.service import ServicePatcher
 from octue.cloud.events.attributes import ResponseAttributes
 from octue.cloud.pub_sub.logging import GoogleCloudPubSubHandler
 from octue.resources.service_backends import GCPPubSubBackend
+from octue.twined.cloud.emulators._pub_sub import MESSAGES, MockService
+from octue.twined.cloud.emulators.service import ServicePatcher
 from tests.base import BaseTestCase
 
 QUESTION_UUID = "96d69278-44ac-4631-aeea-c90fb08a1b2b"
@@ -74,7 +74,7 @@ class TestGoogleCloudPubSubHandler(BaseTestCase):
 
         service = MockService(backend=GCPPubSubBackend(project_id="blah"))
 
-        with patch("octue.cloud.emulators._pub_sub.MockPublisher.publish") as mock_publish:
+        with patch("octue.twined.cloud.emulators._pub_sub.MockPublisher.publish") as mock_publish:
             GoogleCloudPubSubHandler(event_emitter=service._emit_event, attributes=ATTRIBUTES).emit(record)
 
         self.assertEqual(
